+++ src/transactions/transactionBuilder/index.ts
@@ -5,3 +5,4 @@ export * from "./helpers";
 export * from "./transactionBuilder";
 export * from "./remoteAbi";
 export * from "./signingMessage";
+export * from "./structEnumParser";